+++ src/main/java/goormthonuniv/team_22_be/post/application/dto/PostResponseDto.java
@@ -1,4 +1,4 @@
-package goormthonuniv.team_22_be.api.post.dto;
+package goormthonuniv.team_22_be.post.application.dto;
 
 import goormthonuniv.team_22_be.post.domain.model.Post;
 import goormthonuniv.team_22_be.post.domain.model.PostCategory;
