+++ src/main/java/goormthonuniv/team_22_be/post/application/service/PostService.java
@@ -1,8 +1,8 @@
-package goormthonuniv.team_22_be.api.post.service;
+package goormthonuniv.team_22_be.post.application.service;
 
-import goormthonuniv.team_22_be.api.post.dto.PostCreateDto;
-import goormthonuniv.team_22_be.api.post.dto.PostResponseDto;
-import goormthonuniv.team_22_be.api.post.dto.PostUpdateDto;
+import goormthonuniv.team_22_be.post.application.dto.PostCreateDto;
+import goormthonuniv.team_22_be.post.application.dto.PostResponseDto;
+import goormthonuniv.team_22_be.post.application.dto.PostUpdateDto;
 import goormthonuniv.team_22_be.shared.dto.PageResponse;
 import org.springframework.data.domain.Pageable;
