+++ src/main/java/goormthonuniv/team_22_be/post/domain/repository/PostLikeRepository.java
@@ -1,6 +1,6 @@
-package goormthonuniv.team_22_be.api.post.repository;
+package goormthonuniv.team_22_be.post.domain.repository;
 
-import goormthonuniv.team_22_be.api.post.entity.PostLike;
+import goormthonuniv.team_22_be.post.domain.model.PostLike;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostLikeRepository extends JpaRepository<PostLike, Long> {
